Merge pull request #2
This pull request adds the view models 'AccountViewModel' and 'ChatViewModel' classes. The 'AccountViewModel' class contains an 'open_chat_page' function which opens the 'ChatView' and the 'ChatViewModel' contains functions to create the chat boxes for sending and receiving messages. The functions and variables in both the view models have been bounded to the respective views.