Merge pull request #3
This pull request adds the 'App' module. It contains an 'app' function which creates an instance of the 'AccountView' class and calls the 'mainloop' function on it. A thread is created with this function as its target and then the 'start' function is called on the entry point of this module.