Merge pull request #4
This pull request adds the 'Sockets' module with the 'Server' and 'Client' modules for facilitating the transfer of information between two clients. The 'Server' module allows the transfer of information in a one-to-one fashion between the two clients.

The 'Sockets' module is also integrated with the 'ChatViewModel class'. When a 'ChatView' instance is initialized, the client port and target port is prompted through the command line. The send function of the 'sockets' client is bound to the 'typing_button' command, while the 'receive' function from the sockets is delegated to two functions in the 'ChatViewModel' responsible for creating the chat boxes with the message for the sender or the recipient.