Refactor `Account` page to add `Sign Up` functionality
This commit refactors the `AccountView` and `AccountViewModel` to add the `Sign Up` functionality in this app. A static function `get_first_usable_port()` is added to the view model to find the first usable port from the database. Another function `sign_up()` handles the functionality to add a user with a port assigned into the database entry. This function has been bound to the `Sign Up` button in the `AccountView`.