Implement `Sign In` functionality
This commit refactors the `client` module of sockets and the `Account` page to implement the `Sign In` functionality. A function `set_port()` is added to the `client` module which sets the client port of the current user instance. A `sign_in()` function is added to the `AccountViewModel` which retrieves the port assigned to the entered username from the database and then calls the `set_port()` function on the `client` module.