Merge pull request #5
This pull request adds the `Database` package with the class `DatabaseClient` which contains the variables to reference a client instance as well as the database and the `users` collection from the MongoDB database. 

The MongoDB database has been integrated with the application by refactoring the `client` sockets module and the `AccountView` and `AccountViewModel` classes. Now, the user is assigned a port into the local machine directly from the application itself. This behaviour is handled by the `AccountViewModel` during the `sign_up()` and `sign_in()` function calls, which automatically assigns a username a port in the local machine and then retrieves it respectively. The `client` module now contains a function `set_port()` to set the port for one particular client instance.