Add `widgets` package and `UserChatButton` class
This commit adds the class `UserChatButton` in a newly created package `widgets`. This class is a child of the class `CTkButton`. This button will be added as the contacts button in the `Chat` page. It contains the username of the contact as the variable. The target port of the recipient will be received from the database using this username in the class.