Add feature to stop sockets client and remove chatboxes
This commit adds the feature in the `client` socket to run the `receive` function while a global variable `running` is set to true. The value is set to false by calling a function `stop` created inside the `client` socket. The `stop` function is called along with the `exit` function in the `Chat` page whenever the user presses the close button.

The functionality to remove all the chat boxes inside the frame has been added in the `Chat` page whenever the user presses a `UserChatButton` instance inside the Contacts list frame.