Add feature to talk to specific username or target port
This commit removes the `input` function that was run when the thread for receiving messages was started for the client. Instead, a new function `set_target_port()` is added to the `client` socket. This function is called whenever the user presses a `UserChatButton` in the contacts list frame. The port is then received from the database using the `user_name` variable of the `UserChatButton`.