Add `Data` class to add saving feature
This commit adds the class `Data` inside the package `data`. This class has a `chats` variable which is a dictionary to store the contacts and the chats with those contacts. Functions have been added to the data class to save and load the data and also check if the file exists.

The class have been integrated with the app to now save and load the contact names but not the chats.