Merge pull request #7
This pull request adds the feature to save and load the contact and chat data for a user. It adds a class `Data` inside the newly added package `data` which contains a single variable `chats`, a dictionary to save the contact name and the chats list with the contact and functions to save and load that variable from a file and to check if that file exists.

The `Chat` page is then refactored to load the stored contacts as well as the chats from the file. Whenever the user clicks a `UserChatButton`, the corresponding chats are then loaded in the UI for that user.