Add feature to receive multiple and unsent messages
This commit adds the feature in the `client` socket to receive multiple messages from the server, in case the server sends it. It does so by splitting the decoded message into a list of messages and running the newly added function `handle_clients()` for each function, which creates the chat-boxes accordingly.

The `handle_message()` function also checks for the key `unsent` in any of the messages if the message was sent to that client when it was offline. If that key exists, then the chats are loaded into the `chats` variable of the `Data` class and then the `save_data()` function of the `Data` class is run.