Add feature to load chat buttons when unsent message received
This commit adds an event delegate in `client` sockets `RECEIVE_UNSENT_MESSAGE_DELEGATE`, which is run every time an unsent message is sent to the user.

The `ChatViewModel` class is refactored to have a new variable `created_contact_buttons_list` which contains the list of usernames whose corresponding `UserChatButton` has been added. The `load_contacts()` function is refactored to create the buttons of contacts whose buttons have not been created.

The `load_contacts()` function is set as the event delegate `RECEIVE_UNSENT_MESSAGE_DELEGATE` in the `client` sockets.