Add feature to load chats even if user is not selected
This commit adds the feature to load chats from another user if they are sent from a user which is not selected in the UI.

The `client` sockets is refactored to change `RECEIVE_UNSENT_MESSAGE_DELEGATE` to `USER_NOT_SELECTED_MESSAGE_DELEGATE` and another variable `GET_SELECTED_USER_NAME` is created. The `RECEIVE_MESSAGE_DELEGATE_TARGET` is only run if the `GET_SELECTED_USER_NAME()` function returns a username equal to the username which sent the message, otherwise it is just saved to the `Data` class.

The `ChatViewModel` is also refactored to add the function `get_selected_user_name()` which is set to the variable `GET_SELECTED_USER_NAME` in `client` sockets.

So now, a `load_contacts()` function is only run if the username which sent the message is not selected in the UI.