Merge pull request #8
This pull requests add two features, one of the `server` sockets to hold messages sent to usernames which are not connected to the server and then send it back once again once the client becomes connected to the server.
The other feature added is of receiving chats from unknown contacts. Whenever a user receives a chat from an unknown contact, then the corresponding `UserChatButton` is automatically created by the UI.