Add feature to send message with enter key
This commit adds the feature to send the message using the enter key by adding a function `send_message()` in the `ChatView` which checks the message length then runs the `send_message()` function from the `AccountViewModel`.

The enter key release is bound to the removal of the added newline in the textbox.