Add return release key binding to `add_contact_user_name_entry`
This commit adds the functionality inside the `ChatView` to add the contact whenever the user releases the `Return` key when focused on the `add_contact_user_name_entry`. Also, the `<Return>` key binding for `typing_entry` and instead binding the `send_message()` function to the release of `<Return>` key when the user is focused on `typing_entry`.